more int castings

+++ PulseGeneration.py
@@ -225,7 +225,7 @@ def simple_pulse(sampling_rate, params):
 
     total_length = round(duration + params['onset'] + params['offset'], 10) # N.B. Have to round here due to floating point representation problem
 
-    return pulse, np.linspace(0, total_length, total_length*sampling_rate)
+    return pulse, np.linspace(0, total_length, int(total_length*sampling_rate))
 
 
 def multi_simple_pulse(sampling_rate, global_onset, global_offset, params_list):
@@ -302,7 +302,7 @@ def plume_pulse(sampling_rate, params):
     offset = np.zeros(int(sampling_rate * params['offset']))
 
     total_length = round(params['onset'] + params['offset'] + len(pulse) / sampling_rate, 10)
-    return np.hstack((onset, pulse, offset)), np.linspace(0, total_length, total_length * sampling_rate)
+    return np.hstack((onset, pulse, offset)), np.linspace(0, total_length, int(total_length * sampling_rate))
 
 def anti_plume_pulse(sampling_rate, params):
     plume = sio.loadmat(params['data_path'])
@@ -326,7 +326,7 @@ def anti_plume_pulse(sampling_rate, params):
     offset = np.zeros(int(sampling_rate * params['offset']))
 
     total_length = round(params['onset'] + params['offset'] + len(pulse) / sampling_rate, 10)
-    return np.hstack((onset, anti_pulse, offset)), np.linspace(0, total_length, total_length * sampling_rate)
+    return np.hstack((onset, anti_pulse, offset)), np.linspace(0, total_length, int(total_length * sampling_rate))
 
 
 def binary_pulse(sampling_rate, params):
@@ -357,7 +357,7 @@ def binary_pulse(sampling_rate, params):
     offset = np.zeros(int(sampling_rate * params['offset']))
 
     total_length = round(params['onset'] + params['offset'] + len(bin_pulse)/sampling_rate, 10)
-    return np.hstack((onset, bin_pulse, offset)), np.linspace(0, total_length, total_length * sampling_rate)
+    return np.hstack((onset, bin_pulse, offset)), np.linspace(0, total_length, int(total_length * sampling_rate))
 
 
 def dummy_noise_pulse(sampling_rate, params):
@@ -386,7 +386,7 @@ def dummy_noise_pulse(sampling_rate, params):
     offset = np.zeros(sampling_rate * params['offset'])
 
     total_length = round(duration + params['onset'] + params['offset'], 10)
-    return np.hstack((onset, pulse, offset)), np.linspace(0, total_length, total_length * sampling_rate)
+    return np.hstack((onset, pulse, offset)), np.linspace(0, total_length, int(total_length * sampling_rate))
 
 
 def multi_noise_pulse(sampling_rate, global_onset, global_offset, params_list):
